# Improve/fix Cat performance