--- a/benchmark/bench_cat.jl
+++ b/benchmark/bench_cat.jl
@@ -1,0 +1,14 @@
+module BenchCat
+
+using BenchmarkTools
+using Transducers
+
+itr = (y for x in 1:1000 for y in 1:x);
+
+suite = BenchmarkGroup()
+
+suite["xf"] = @benchmarkable foldl(+, eduction($itr))
+suite["base"] = @benchmarkable sum($itr)
+
+end  # module
+BenchCat.suite